requirements.txt
mediapipe
opencv-python
scikit-learn
pandas
numpy
seaborn
tk
Pillow